backend/requirements.txt
flask
python-dotenv
gunicorn
flask-sqlalchemy
pymysql
